transcribe_website/transcriber_backend/pyproject.toml
[project]
name = "src"
version = "0.1.0"
requires-python = ">=3.10, <3.14"
dependencies = [
    # Better datetime
    "arrow>=1.3.0",
    # Translator
    "deep-translator>=1.11.4",
    "deepmultilingualpunctuation>=1.0.1",
    "faster-whisper==1.0.1",
    # Display sizes in human readable format
    "humanize>=4.11.0",
    "langdetect>=1.0.9",
    # Had to fix version to make poetry work
    "llvmlite>=0.43.0",
    # Logging
    "loguru>=0.7.2",
    "numba>=0.60.0",
    "numpy<2",
    # Database interaction
    "prisma>=0.15.0",
    # Better dataclasses
    "pydantic>=2.10.1",
    "python-dotenv>=1.0.1",
    # Argparse with dataclass
    "simple-parsing>=0.1.6",
    # https://stackoverflow.com/a/76359919/10882657
    "torch>=2.4.0",
    "torchaudio>=2.4.0",
    # Progress bar
    "tqdm>=4.67.1",
    "whisperx>=3.1.6",
]

[tool.uv]
dev-dependencies = [
    # Type checker
    "pyre-check>=0.9.23",
    # Linter
    "ruff>=0.8.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchaudio = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.yapf]
based_on_style = "pep8"
column_limit = 120
split_arguments_when_comma_terminated = true
dedent_closing_brackets = true
allow_split_before_dict_value = false

[tool.ruff]
target-version = 'py310'
line-length = 120

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
select = [
    "C4",  # flake8-comprehensions
    "E",   # Error
    "F",   # pyflakes
    "BLE", # flake8-blind-except
    # "I",   # isort
    "N",   # pep8-naming
    "PGH", # pygrep-hooks
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "W",   # Warning
    "Q",   # flake8-quotes
    "YTT", # flake8-2020
    "UP",  # pyupgrade
    #    "A",  # flake8-builtins
]
# Allow Pydantic's `@validator` decorator to trigger class method treatment.
pep8-naming.classmethod-decorators = ["pydantic.validator", "classmethod"]
